Move `ThemeTextField` to its own file
